# pozoristeback/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Teatar Maska API",
        "description": "API dokumentacija za seminarski rad",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Lokalni server"
        }
    ],
    "paths": {
...
        "/api/predstave/{predstavaId}/izvodjenja": {
            "get": {
                "tags": [
                    "Izvođenja"
                ],
                "summary": "Sva izvođenja određene predstave",
                "description": "Vraća sva izvođenja za određenu predstavu (filtriranje).",
                "operationId": "f3b0f7ae56419cbb1ecdd9dbd22e4870",
                "parameters": [
                    {
                        "name": "predstavaId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešno"
                    }
                }
            }
        },
...
        "/api/predstave": {
            "get": {
                "tags": [
                    "Predstave"
                ],
                "summary": "Prikaz svih predstava",
                "operationId": "bd75bc54dfef10a9df4d727d289388b0",
                "responses": {
                    "200": {
                        "description": "Uspešno učitane predstave"
                    }
                }
            },
            "post": {
                "tags": [
                    "Predstave"
                ],
                "summary": "Dodavanje nove predstave",
                "description": "Store a newly created resource in storage.",
                "operationId": "866e48ac223988079a5f535acb5fde5b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "naziv",
                                    "opis",
                                    "reditelj",
                                    "trajanje_min",
                                    "img_url"
                                ],
                                "properties": {
                                    "naziv": {
                                        "type": "string",
                                        "example": "Gospođa ministarka"
                                    },
                                    "opis": {
                                        "type": "string",
                                        "example": "Komedija Branislava Nušića"
                                    },
                                    "reditelj": {
                                        "type": "string",
                                        "example": "Jagoš Marković"
                                    },
                                    "trajanje_min": {
                                        "type": "integer",
                                        "example": 105
                                    },
                                    "img_url": {
                                        "type": "string",
                                        "example": "https://link-ka-slici.com/slika.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Uspešno kreirano"
                    },
                    "422": {
                        "description": "Greška u validaciji"
                    }
                }
            }
        },
        "/api/predstave/{id}": {
            "get": {
                "tags": [
                    "Predstave"
                ],
                "summary": "Prikaz jedne predstave",
                "description": "Display the specified resource.",
                "operationId": "62b55ce8585f385297dbdc4b9c886a68",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešno"
                    },
                    "404": {
                        "description": "Nije pronađena"
                    }
                }
            },
            "put": {
                "tags": [
                    "Predstave"
                ],
                "summary": "Ažuriranje postojeće predstave",
                "description": "Update the specified resource in storage.",
                "operationId": "db9b9c245a38ffd1e67ac40627fb0dd2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID predstave koju ažurirate",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "naziv": {
                                        "type": "string",
                                        "example": "Nova verzija naziva"
                                    },
                                    "opis": {
                                        "type": "string",
                                        "example": "Novi opis predstave..."
                                    },
                                    "reditelj": {
                                        "type": "string",
                                        "example": "Ime novog reditelja"
                                    },
                                    "trajanje_min": {
                                        "type": "integer",
                                        "example": 110
                                    },
                                    "img_url": {
                                        "type": "string",
                                        "example": "https://novi-link.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Predstava uspešno ažurirana"
                    },
                    "404": {
                        "description": "Predstava nije pronađena"
                    },
                    "422": {
                        "description": "Greška u validaciji podataka"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Predstave"
                ],
                "summary": "Brisanje predstave",
                "description": "Remove the specified resource from storage.",
                "operationId": "ee740461ad31c4379ff42f03ca5d59d9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Obrisano"
                    }
                }
            }
        },
        "/api/predstave/pretraga": {
            "get": {
                "tags": [
                    "Predstave"
                ],
                "summary": "Pretraga predstava po nazivu",
                "operationId": "0634ec242ad2bd5739d3bc78827a5b7d",
                "parameters": [
                    {
                        "name": "naziv",
                        "in": "query",
                        "description": "Deo naziva predstave za pretragu",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešna pretraga"
                    },
                    "400": {
                        "description": "Loš zahtev (fali parametar)"
                    }
                }
            }
        },
        "/api/predstave/{id}/izvodjenja": {
            "get": {
                "tags": [
                    "Predstave"
                ],
                "summary": "Prikaz svih izvođenja za određenu predstavu",
                "operationId": "7c2a31d9511f3fc632866ce97061b7a1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID predstave",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešno učitana izvođenja"
                    },
                    "404": {
                        "description": "Predstava nije pronađena"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Unesite vaš API token koji ste dobili pri prijavi (login) ili registraciji.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}